add fetch embed util
